requirements.txt
dj-database-url==0.5.0
Django==2.1.2
gunicorn==19.9.0
whitenoise==4.1.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
django-heroku==0.3.1
django-tinymce4-lite==1.7.4